+++ asyncgit/src/sync/remotes/mod.rs
@@ -51,6 +51,81 @@ pub fn get_default_remote(repo_path: &RepoPath) -> Result<String> {
 	get_default_remote_in_repo(&repo)
 }
 
+/// Gets the current branch the user is on.
+/// Returns none if they are not on a branch
+/// and Err if there was a problem finding the branch
+fn get_current_branch(
+	repo: &Repository,
+) -> Result<Option<git2::Branch>> {
+	for b in repo.branches(None)? {
+		let branch = b?.0;
+		if branch.is_head() {
+			return Ok(Some(branch));
+		}
+	}
+	Ok(None)
+}
+
+/// Tries to find the default repo to push to based on configuration.
+///
+/// > remote.pushDefault
+/// >
+/// > The remote to push to by default. Overrides `branch.<name>.remote` for all branches, and is
+/// > overridden by `branch.<name>.pushRemote` for specific branches.
+///
+/// > branch.<name>.remote
+/// >
+/// > When on branch `<name>`, it tells `git fetch` and `git push` which remote to fetch from or
+/// > push to. The remote to push to may be overridden with `remote.pushDefault` (for all
+/// > branches). The remote to push to, for the current branch, may be further overridden by
+/// > `branch.<name>.pushRemote`. If no remote is configured, or if you are not on any branch and
+/// > there is more than one remote defined in the repository, it defaults to `origin` for fetching
+/// > and `remote.pushDefault` for pushing.
+///
+/// [git-config-remote-push-default]: https://git-scm.com/docs/git-config#Documentation/git-config.txt-remotepushDefault
+/// [git-config-branch-name-remote]: https://git-scm.com/docs/git-config#Documentation/git-config.txt-branchltnamegtremote
+///
+/// Falls back to `get_default_remote_in_repo`.
+pub fn get_default_remote_for_push(
+	repo_path: &RepoPath,
+) -> Result<String> {
+	let repo = repo(repo_path)?;
+	get_default_remote_for_push_in_repo(&repo)
+}
+
+pub(crate) fn get_default_remote_for_push_in_repo(
+	repo: &Repository,
+) -> Result<String> {
+	scope_time!("get_default_remote_for_push_in_repo");
+
+	let config = repo.config()?;
+
+	let branch = get_current_branch(repo)?;
+
+	if let Some(branch) = branch {
+		let remote_name = bytes2string(branch.name_bytes()?)?;
+
+		let entry_name =
+			format!("branch.{}.pushRemote", &remote_name);
+
+		if let Ok(entry) = config.get_entry(&entry_name) {
+			return bytes2string(entry.value_bytes());
+		}
+
+		if let Ok(entry) = config.get_entry("remote.pushDefault") {
+			return bytes2string(entry.value_bytes());
+		}
+
+		let entry_name = format!("branch.{}.remote", &remote_name);
+
+		if let Ok(entry) = config.get_entry(&entry_name) {
+			return bytes2string(entry.value_bytes());
+		}
+	}
+
+	get_default_remote_in_repo(repo)
+}
+
 /// see `get_default_remote`
 pub(crate) fn get_default_remote_in_repo(
 	repo: &Repository,
@@ -299,9 +374,68 @@ mod tests {
 			]
 		);
 
-		let res =
+		let default_remote =
 			get_default_remote_in_repo(&repo(repo_path).unwrap());
-		assert_eq!(res.is_err(), true);
-		assert!(matches!(res, Err(Error::NoDefaultRemoteFound)));
+
+		assert!(matches!(
+			default_remote,
+			Err(Error::NoDefaultRemoteFound)
+		));
+	}
+
+	#[test]
+	fn test_default_remote_for_push() {
+		let (remote_dir, _remote) = repo_init().unwrap();
+		let remote_path = remote_dir.path().to_str().unwrap();
+		let (repo_dir, repo) = repo_clone(remote_path).unwrap();
+		let repo_path: &RepoPath = &repo_dir
+			.into_path()
+			.as_os_str()
+			.to_str()
+			.unwrap()
+			.into();
+
+		debug_cmd_print(
+			repo_path,
+			"git remote rename origin alternate",
+		);
+
+		debug_cmd_print(
+			repo_path,
+			&format!("git remote add someremote {remote_path}")[..],
+		);
+
+		let mut config = repo.config().unwrap();
+
+		config
+			.set_str("branch.master.remote", "branchremote")
+			.unwrap();
+
+		let default_push_remote =
+			get_default_remote_for_push_in_repo(&repo);
+
+		assert!(
+			matches!(default_push_remote, Ok(remote_name) if remote_name == "branchremote")
+		);
+
+		config.set_str("remote.pushDefault", "pushdefault").unwrap();
+
+		let default_push_remote =
+			get_default_remote_for_push_in_repo(&repo);
+
+		assert!(
+			matches!(default_push_remote, Ok(remote_name) if remote_name == "pushdefault")
+		);
+
+		config
+			.set_str("branch.master.pushRemote", "branchpushremote")
+			.unwrap();
+
+		let default_push_remote =
+			get_default_remote_for_push_in_repo(&repo);
+
+		assert!(
+			matches!(default_push_remote, Ok(remote_name) if remote_name == "branchpushremote")
+		);
 	}
 }